Extract simple EventTile adapters
